v0.9.0.0
Major updates. Removed the stuff dealing with the Alturos.Yolo wrapper and put it in igorCore. Now, the igorCore can be used for different interfaces (hint: possible GUI coming).

Swapped to a pending pull request of Alturos.Yolo that is more dynamic about detecting VC++ Redistributable and seems to work nicely.

Also, working with an actual git pull of the Alturos.Yolo project, rather than just a locally saved download of it.